requirements.txt
ipython==8.15.0
ipywidgets==8.1.1
openai==1.3.7
python-dotenv==1.0.0